examples/redteam/crypto_stealing_demo/pyproject.toml
[project]
name = "red-teaming-demo"
version = "0.1.0"
description = "Demo package for the Red Teaming AI Agents talk at AI Builders Amsterdam"
requires-python = ">=3.12"
dependencies = [
  "evaluatorq[redteam]>=1.3.0rc2",
  "fastapi>=0.115",
  "sse-starlette>=2.1",
  "uvicorn[standard]>=0.32",
  "rich>=13.9",
  "httpx>=0.27",
  "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "ruff>=0.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py312"
